chore(validation): extract web-cookie, kiro, specialty-inline validators into validation/* leaves
