+++ src/lib/providers/validation/webCookie.ts
@@ -1,0 +1,152 @@
+// Web-cookie session-ping validator + Bytez auth-only probe. Extracted from validation.ts
+// (god-file decomposition) — top-level functions with no dispatcher-state captures; behavior is
+// byte-identical to the original inline defs.
+import { WEB_COOKIE_PROVIDERS, isLocalProvider } from "@/shared/constants/providers";
+import { getRegistryEntry } from "@omniroute/open-sse/config/providerRegistry.ts";
+import { normalizeBaseUrl } from "./urlHelpers";
+import { STANDARD_USER_AGENT, buildBearerHeaders } from "./headers";
+import {
+  validationRead,
+  toValidationErrorResult,
+  toWebCookieValidationErrorResult,
+  WEB_COOKIE_PROVIDERS_WITHOUT_MODELS_API,
+} from "./transport";
+
+/**
+ * Validates web-cookie providers by performing a ping request to check if the session is still valid.
+ * Returns SESSION_EXPIRED error code if the upstream returns 401/403.
+ */
+export async function validateWebCookieProvider({
+  provider,
+  apiKey,
+  providerSpecificData: _providerSpecificData = {},
+}: {
+  provider: string;
+  apiKey?: string;
+  providerSpecificData?: Record<string, unknown>;
+}) {
+  try {
+    const entry = getRegistryEntry(provider);
+    const cookieProvider = WEB_COOKIE_PROVIDERS[provider as keyof typeof WEB_COOKIE_PROVIDERS];
+    if (!entry && !cookieProvider) {
+      return { valid: false, error: "Provider not found in registry", unsupported: true };
+    }
+
+    // For web-cookie providers, apiKey contains the cookie string
+    const cookie = (apiKey || "").trim();
+    if (!cookie) {
+      return { valid: false, error: "Cookie required for web-cookie provider", unsupported: false };
+    }
+
+    if (!entry) {
+      // Providers listed in WEB_COOKIE_PROVIDERS without a providerRegistry entry (e.g.
+      // gemini-business, poe-web, venice-web, v0-vercel-web) only expose a
+      // marketing website URL, not a real API host. Probing `${website}/models`
+      // does not reliably signal session validity for these —
+      // live verification showed most return redirects or SPA 200s regardless of
+      // cookie validity, which would silently report an expired/garbage cookie as
+      // "OK" (worse than an honest "not supported"). Until each of these providers
+      // has a verified, side-effect-free auth probe against its real API host, report
+      // unsupported instead of a false positive.
+      return {
+        valid: false,
+        error: "Provider validation not supported",
+        unsupported: true,
+      };
+    }
+
+    // Attempt a minimal request to check if the session is valid
+    // Use /models endpoint or a minimal completion request depending on the provider
+    const baseUrl = normalizeBaseUrl(entry.baseUrl || "");
+
+    // Defense-in-depth: only an http(s) baseUrl without a query string is safe to
+    // probe by blindly appending `/models`. A ws(s):// baseUrl (e.g. copilot-web) is
+    // already rejected by the outbound URL guard downstream, but reject it explicitly
+    // here for the honest "unsupported" result instead of a confusing security-block
+    // message — this also covers a future http(s) baseUrl carrying a query string,
+    // which the guard does not currently block (#7857 acceptance criteria).
+    if (!/^https?:\/\//i.test(baseUrl) || baseUrl.includes("?")) {
+      return {
+        valid: false,
+        error: "Provider validation not supported",
+        unsupported: true,
+      };
+    }
+
+    const testUrl = `${baseUrl}/models`;
+
+    const res = await validationRead(
+      testUrl,
+      {
+        method: "GET",
+        headers: {
+          "User-Agent": STANDARD_USER_AGENT,
+          Cookie: cookie,
+        },
+      },
+      isLocalProvider(provider)
+    );
+
+    if (res.status === 401 || res.status === 403) {
+      return {
+        valid: false,
+        error: "SESSION_EXPIRED",
+        errorCode: "AUTH_007",
+        unsupported: false,
+      };
+    }
+
+    // #7857: for providers whose baseUrl is a conversation/completion endpoint rather
+    // than a real API root, the /models path never existed upstream — a redirect,
+    // login-HTML 200, 404, 405, or 429 from it is not a meaningful auth signal and is
+    // indistinguishable from a genuinely valid session. Report the same honest
+    // "unsupported" result the !entry branch above already gives its no-registry
+    // siblings, instead of a false `valid: true`.
+    if (WEB_COOKIE_PROVIDERS_WITHOUT_MODELS_API.has(provider)) {
+      return {
+        valid: false,
+        error: "Provider validation not supported",
+        unsupported: true,
+      };
+    }
+
+    // Any other response (200, 404, 405, 429, ...) means the cookie was accepted —
+    // a 401/403 from the /models probe is the only definitive "session expired" signal
+    // for web-cookie auth, so a non-auth status is treated as a valid session.
+    return { valid: true, error: null, unsupported: false };
+  } catch (error: unknown) {
+    return toWebCookieValidationErrorResult(provider, error);
+  }
+}
+
+// #5422: Bytez key validation cannot use a chat probe. A Bytez account only serves models
+// that have been added to its catalog, so even Bytez's own documented model ids return 404
+// ("Model does not exist or has yet to be added to the Bytez catalog") for a fresh/free key —
+// the generic OpenAI-like chat probe misreads that 404 as "endpoint not supported". Validate
+// against the model-independent, auth-only tasks endpoint instead (verified live):
+//   GET …/models/v2/list/tasks → 200 (valid key) | 401 { error: "Unauthorized" } (invalid).
+// The pure status→result mapping is factored out so it is unit-testable without network.
+export function bytezValidationResultFromStatus(status: number): {
+  valid: boolean;
+  error: string | null;
+} {
+  if (status === 200) {
+    return { valid: true, error: null };
+  }
+  if (status === 401 || status === 403) {
+    return { valid: false, error: "Invalid API key" };
+  }
+  return { valid: false, error: `Validation failed: ${status}` };
+}
+
+export async function validateBytezProvider({ apiKey, providerSpecificData = {} }: any) {
+  try {
+    const res = await validationRead("https://api.bytez.com/models/v2/list/tasks", {
+      method: "GET",
+      headers: buildBearerHeaders(apiKey, providerSpecificData),
+    });
+    return bytezValidationResultFromStatus(res.status);
+  } catch (error: unknown) {
+    return toValidationErrorResult(error);
+  }
+}
